# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ERP MIF Maroc - Development",
    "description": "Backend ERP pour la gestion des interventions industrielles",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents/": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Uploader un document",
        "description": "Upload d’un fichier lié à une intervention (photo, rapport, preuve, etc.).",
        "operationId": "upload_document_api_v1_documents__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Intervention Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Lister tous les documents",
        "description": "Accès à la liste de tous les documents liés aux interventions (admin/responsable uniquement).",
        "operationId": "list_documents_api_v1_documents__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Api V1 Documents  Get"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          }
        }
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Uploader un document (alias)",
        "operationId": "upload_document_alias_api_v1_documents_upload_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Intervention Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_alias_api_v1_documents_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{intervention_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Lister les documents d'une intervention",
        "operationId": "list_documents_by_intervention_api_v1_documents__intervention_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Intervention Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents By Intervention Api V1 Documents  Intervention Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Supprimer un document",
        "description": "Supprime le document (enregistrement + fichier sur disque)",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Uploader un document",
        "description": "Upload d’un fichier lié à une intervention (photo, rapport, preuve, etc.).",
        "operationId": "upload_document_documents__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Intervention Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_documents__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Lister tous les documents",
        "description": "Accès à la liste de tous les documents liés aux interventions (admin/responsable uniquement).",
        "operationId": "list_documents_documents__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Documents  Get"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          }
        }
      }
    },
    "/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Uploader un document (alias)",
        "operationId": "upload_document_alias_documents_upload_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Intervention Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_alias_documents_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{intervention_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Lister les documents d'une intervention",
        "operationId": "list_documents_by_intervention_documents__intervention_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "intervention_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Intervention Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents By Intervention Documents  Intervention Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Supprimer un document",
        "description": "Supprime le document (enregistrement + fichier sur disque)",
        "operationId": "delete_document_documents__document_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_alias_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_alias_api_v1_documents_upload_post"
      },
      "Body_upload_document_alias_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_alias_documents_upload_post"
      },
      "Body_upload_document_api_v1_documents__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents__post"
      },
      "Body_upload_document_documents__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_documents__post"
      },
...
      "DocumentOut": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date_upload": {
            "type": "string",
            "format": "date-time",
            "title": "Date Upload"
          },
          "intervention_id": {
            "type": "integer",
            "title": "Intervention Id"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "path",
          "id",
          "date_upload",
          "intervention_id"
        ],
        "title": "DocumentOut",
        "description": "Schéma renvoyé par l'API pour un document :\n- Contient les métadonnées complètes"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}